feat : 소수 계산 로직 추가

+++ src/main/java/calculator/CalculatorController.java
@@ -9,7 +9,7 @@ public class CalculatorController {
 
     public void calculate() {
         String stringExpression = View.readStringExpression();
-        int result = calculator.calculate(stringExpression);
+        Double result = calculator.calculate(stringExpression);
 
         View.writeResult(result);
     }
